requirements.txt
python-a2a
fastmcp
langchain-openai
yfinance
pandas
requests
beautifulsoup4
python-dotenv
flask